refactor(telegram): extract the Bot API transport into telegram-api.ts
telegram.ts fused three layers: the channel handler (verify/route), the live-stream
renderer, and the Telegram Bot API wire calls. Split the wire layer into
telegram-api.ts (no agent, no rendering): callBotApi (429), sendMessage (4096-split +
HTML→plain fallback), sendMessageDraft, resolveBotUsername, and the file download
path. The channel now imports those and keeps policy + streaming.

This also kills the getFile→size-check→download→byte-check duplication that
fetchTelegramImage and downloadTelegramFile each carried — both now share one
getFileBytes. telegram.ts drops 602→427 lines. No public surface or behavior change
(the api internals were never exported; tests assert via the fetch mock, unchanged).
